deleting user entries whenm user is deleted

+++ src/main/java/pl/umk/mat/danielsz/journal/services/implementations/UserServiceImpl.java
@@ -72,6 +72,17 @@ public class UserServiceImpl implements UserService {
 
     @Override
     public ResponseEntity<Void> deleteOne(String id) {
+        if(id == null){
+            throw new OperationNotAllowedException("Users id cannot be null!");
+        }
+
+        if(!existsById(id)){
+            throw new OperationNotAllowedException("User doesn't exists!");
+        }
+
+        User user = findOne(id);
+
+        entryService.deleteAll(user.getEntries());
         userRepository.deleteById(id);
 
         return new ResponseEntity<Void>(HttpStatus.NO_CONTENT);
